+++ src/client/app/epics/index.ts
@@ -1,0 +1,9 @@
+import { combineEpics } from 'redux-observable';
+
+import { fetchTodosEpic } from './todos';
+
+const rootEpic = combineEpics(
+  fetchTodosEpic,
+);
+
+export default rootEpic;